refactor: moved form logic to form component and directly rendered Form (got rid of EditRecipe component) docs: added some comments throughout
