+++ frontend/src/pages/Recipes.tsx
@@ -2,11 +2,7 @@ import Card from "../components/Card";
 import { useLoaderData, Link } from "react-router-dom";
 
 function Recipes() {
-  // removed useEffect because useEffect should be used for side effects
-  //such as when a certain param changes and not for lifecycle methods
-  //useEffect is a hook in React that allows us to run side effects or
-  //manage state changes after the component has rendered and the DOM has been updated.
-  // instead we use useloaderdata
+  // removed useEffect because useEffect because I already load data with the react-router data api
   const recipes = useLoaderData() as RecipeType[];
   return (
     <div className="m-[50px]">
